requirements.txt
# 核心依赖
streamlit>=1.28.0
docker>=6.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0  # 用于 RANSAC（几何校正）

# 开发依赖
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
ruff>=0.1.0

# 类型检查（可选）
mypy>=1.5.0
types-requests>=2.31.0

